+++ tests/testthat/test-snapshot.R
@@ -15,7 +15,6 @@ test_that("snapshot is idempotent", {
 
 test_that("snapshot failures are reported", {
 
-  renv_scope_envvars(RENV_PATHS_ROOT = renv_scope_tempfile())
   renv_tests_scope("oatmeal")
   init()
 
@@ -441,6 +440,17 @@ test_that("snapshot always reports on R version changes", {
   })
 })
 
+test_that("user can choose to install missing packages", {
+
+  # use a temporary cache to guarantee packages are fully installed
+  # regardless of order other tests are run in
+  renv_scope_envvars(RENV_PATHS_CACHE = renv_scope_tempfile("renv-tempcache-"))
+
+  renv_tests_scope("egg")
+  renv_scope_options(renv.menu.choice = 2)
+  expect_snapshot(snapshot())
+
+})
 
 test_that("useful error message if implicit dep discovery is slow", {